chore: reorganize code structure and add uni-tests
